requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
authlib==1.6.11
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpickle==3.1.2
cryptography==46.0.7
distro==1.9.0
docstring-parser==0.18.0
dotenv==0.9.9
fastapi==0.135.3
frozenlist==1.8.0
google-adk==1.30.0
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-cloud-aiplatform==1.148.0
google-cloud-appengine-logging==1.9.0
google-cloud-audit-log==0.5.0
google-cloud-bigquery==3.41.0
google-cloud-bigquery-storage==2.37.0
google-cloud-bigtable==2.36.0
google-cloud-core==2.5.1
google-cloud-dataplex==2.18.0
google-cloud-discoveryengine==0.13.12
google-cloud-iam==2.22.0
google-cloud-logging==3.15.0
google-cloud-monitoring==2.30.0
google-cloud-pubsub==2.37.0
google-cloud-resource-manager==1.17.0
google-cloud-secret-manager==2.27.0
google-cloud-spanner==3.65.0
google-cloud-speech==2.38.0
google-cloud-storage==3.10.1
google-cloud-trace==1.19.0
google-crc32c==1.8.0
google-genai==1.73.1
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
graphviz==0.21
grpc-google-iam-v1==0.14.4
grpc-interceptor==0.15.4
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib-metadata==8.7.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mako==1.3.11
markupsafe==3.0.3
mcp==1.27.0
mmh3==5.2.1
multidict==6.7.1
opentelemetry-api==1.38.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==26.1
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.13.1
pydantic-core==2.46.1
pydantic-settings==2.13.1
pygments==2.20.0
pyjwt==2.12.1
pyopenssl==26.0.0
pyparsing==3.3.2
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
pyyaml==6.0.3
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.49
sqlalchemy-spanner==1.17.3
sqlparse==0.5.5
sse-starlette==3.3.4
starlette==0.52.1
tenacity==9.1.4
typing-extensions==4.15.0
typing-inspection==0.4.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.44.0
watchdog==6.0.0
websockets==15.0.1
yarl==1.23.0
zipp==3.23.1
